requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Jinja2==3.1.2
mysql-connector-python==8.0.32
pandas==2.0.0
python-dateutil==2.8.2
pytest==7.3.0
werkzeug==2.2.3
sqlalchemy~=1.4.47
faker==18.4.0
flask_restx
newpy~=1.1.16
gunicorn
requests~=2.28.2
flask-cors==3.0.10
flasgger~=0.9.5
flask_caching
PyJWT==2.6.0